Refactor so that basic architecture of the MoE and ViT models is shared.
